removed bugs and warnings for arm m3-m4 target

+++ src/api/share.h
@@ -13,8 +13,8 @@ extern "C" {
 extern uint8_t mutex_lock(uint32_t *mutex);
 extern void mutex_unlock(uint32_t *mutex);
 extern uint8_t spin_lock(uint32_t *mutex, uint16_t retry);
-extern uint8_t semaphore_increment(uint8_t *semaphore, uint16_t retry);
-extern uint8_t semaphore_decrement(uint8_t *semaphore, uint16_t retry);
+extern uint8_t semaphore_increment(uint32_t *semaphore, uint16_t retry);
+extern uint8_t semaphore_decrement(uint32_t *semaphore, uint16_t retry);
 
 #ifdef	__cplusplus
 }
